requirements.txt
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
black==25.11.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.1.8
coloredlogs==15.0.1
cryptography==46.0.3
cssselect==1.3.0
Deprecated==1.3.1
deprecation==2.1.0
distlib==0.4.0
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
faiss-cpu==1.13.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
flake8==7.3.0
flatbuffers==25.9.23
fsspec==2025.10.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
img2pdf==0.6.3
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
isort==6.1.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.9.2
nodeenv==1.9.1
numpy==2.0.2
oauthlib==3.3.1
ocrmypdf==15.4.4
onnxruntime==1.19.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pdfminer.six==20251107
pikepdf==9.11.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
posthog==5.4.0
pre_commit==4.3.0
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.8.0
tqdm==4.67.1
transformers==4.57.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
zipp==3.23.0
lxml-html-clean==0.4.3
openai
chromadb
scikit-learn
nltk
python-dotenv
tqdm